MDL-83297 gradereport_grader: Lower MAX_GRADES_PER_PAGE constant
After deploying to our production environment, we noticed that for large gradebooks, having a max of 200,000 grades per
page was still too high and resulted in the gradebook page crashing or failing to load entirely.  We tested different values
of if and concluded that 70,000 was roughly the maximum for which the gradebook works with a satisfactory performance.